requirements_antonio.txt
absl-py==2.3.1
ale-py==0.11.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.4.0
Bottleneck
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
grpcio==1.76.0
gymnasium==1.1.1
idna==3.11
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
lz4==4.4.5
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
metadrive-simulator @ git+https://github.com/metadriverse/metadrive.git@85e5dadc6c7436d324348f6e3d8f8e680c06b4db
mkl-service==2.5.2
mkl_fft
mkl_random
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numexpr
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
ormsgpack==1.7.0
Panda3D==1.10.15
panda3d-gltf==0.15.0
panda3d-simplepbr==0.13.1
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow
platformdirs==4.4.0
progressbar==2.5
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.6.1
Pygments==2.19.2
pyparsing
python-dateutil
pytz
PyVirtualDisplay==3.0
PyYAML==6.0.3
ray==2.51.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
six
stable_baselines3==2.7.0
stack-data==0.6.3
sympy==1.14.0
tdqm==0.0.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
yapf==0.43.0
